[native] Move sendTextMessage to ChatInputStateContainer
